grpc-python-server/requirements.txt
grpcio-tools==1.25.0
grpcio==1.25.0
protobuf==3.11.1
requests==2.22.0
